@ricardoV94 2024-11-29
Here is a direct comparison of C and numba backends for the non C-contiguous case:

```python
import numpy as np
import pytensor

c_contiguous = False
for transpose_in_graph in (True, False):
    rng = np.random.default_rng(123)
    N = 256
    x_test = rng.uniform(size=(N, N, N))
    transpose_axis = (0, 1, 2) if c_contiguous else (2, 0, 1)
    
    if not transpose_in_graph:
        x_test = x_test.transpose(transpose_axis)
    
    x = pytensor.shared(x_test, name="x", shape=x_test.shape, borrow=True)
    
    if transpose_in_graph:
        x = x.transpose(transpose_axis)
        
    out = x.sum(axis=0)
    c_fn = pytensor.function([], out, mode="FAST_COMPILE")
    numba_fn = pytensor.function([], out, mode="NUMBA")
    np.testing.assert_allclose(c_fn(), numba_fn())
    print(f"{transpose_in_graph=}")
    %timeit c_fn()
    %timeit numba_fn()
          
# transpose_in_graph=True
# 33.7 ms ± 2.25 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
# 188 ms ± 4.05 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
# transpose_in_graph=False
# 33 ms ± 1.15 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
# 103 ms ± 1.96 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

Surprisingly transposing in the graph is very slow in itself. Perhaps our implementation of `DimShuffle` is doing a useless copy? I see a Reshape which in numba requires a `np.ascontiguousarray(x)`. But we shouldn't need any reshape in this graph.

Now a direct numba implementation is much faster, although I think it's exactly the same code as in this PR??

```python
import numpy as np
import numba

c_contiguous = False
rng = np.random.default_rng(123)
N = 256
x_test = rng.uniform(size=(N, N, N))
transpose_axis = (0, 1, 2) if c_contiguous else (2, 0, 1)
x = x_test.transpose(transpose_axis)

@numba.njit(fastmath=True, boundscheck=False)
def numba_fn(x):
    x_shape = x.shape
    out_shape = (x_shape[1], x_shape[2])
    out = np.full(out_shape, np.asarray(0.0).item(), dtype="float64")
    for i0 in range(x_shape[0]):
        for i1 in range(x_shape[1]):
            for i2 in range(x_shape[2]):
                out[i1, i2] += x[i0, i1, i2]
    return out

np.testing.assert_allclose(numba_fn(x_test), np.sum(x_test, 0))
%timeit numba_fn(x_test)  
# 8.04 ms ± 461 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```