@ricardoV94 2024-11-29
Numba doing badly on the non-contiguous case is all due to loop ordering. LLVM doesn't reorder based on strides :(

Anyway this PR improves overall, better old speeds where just due to luck when the reduced loop was the fastest